The Area and Premiers
The table shows the area and premier from the tumor and non-tumor annotated regions. The first column shows the name of WSIs. The second column shows the region_id, which can be a tumor or non-tumor region. The third column shows the region names. The fourth and fifth columns show the area and premier of the regions.